pyproject.toml
[project]
name = "home-task-wargaming"
version = "0.1.0"
description = "Home task for Wargaming"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "N",  # pep8-naming
    "FIX",
    "FBT",
    "ARG",
    "PTH",
    "ERA",
    "RUF",
]
ignore = ["RUF012"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
line-ending = "lf"

